pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hexastorm"
version = "0.1.0"
description = "amaranth framework for writing to a substrate with a laser"
authors = [
    { name = "Rik Starmans", email = "hstarmans@hexastorm.com" }
]
license = { file = "LICENSE.txt" }
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "amaranth-boards",
    "cairosvg>=2.7.1",
    "fire>=0.7.1",
    "numba>=0.60.0",
    "numpy>=2.0.0",
    "opencv-python>=4.10.0",
    "pillow>=11.2.1",
    "tqdm>=4.67.3",
    "yowasp-nextpnr-ice40>=0.8.0.0.post621",
    "yowasp-yosys>=0.53.0.0.post912",
]

[dependency-groups]
dev = [
    "nbqa>=1.9.1",
    "pytest>=8.4.2",
    "ruff>=0.11.12",
]
plot = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.9.0",
]

[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "double"

[tool.uv.sources]
amaranth-boards = { git = "https://github.com/amaranth-lang/amaranth-boards.git", rev = "c26a72e59c786b38e0e989ae64c6c2560ca7c29c" }
